requirements.txt
boto3==1.28.78
botocore==1.31.78
certifi==2023.7.22
cffi @ file:///private/tmp/cffi-20231014-5216-8jirg3/cffi-1.16.0
charset-normalizer==3.3.2
coinbase==2.1.0
docopt==0.6.2
geminipy==0.0.4
idna==3.4
install==1.3.5
jmespath==1.0.1
mido==1.3.0
packaging @ file:///private/tmp/python-packaging-20231018-5109-frjsuy/packaging-23.2
pipreqs==0.4.13
pycparser @ file:///private/tmp/pycparser-20231013-5387-1wj1ewr/pycparser-2.21
pycryptodome==3.19.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-rtmidi==1.5.7
regex==2023.10.3
requests==2.31.0
rtmidi==2.5.0
s3transfer==0.7.0
semantic-version==2.10.0
setuptools-rust==1.8.1
six==1.16.0
unicodedata2==15.1.0
urllib3==2.0.7
yarg==0.1.9
